rename echomw
