+++ internal/httpserve/middleware/echocontext/context.go
@@ -1,4 +1,4 @@
-package echomw
+package echocontext
 
 import (
 	"context"
