+++ internal/httpserve/middleware/echocontext/doc.go
@@ -1,0 +1,2 @@
+// Package echocontext adds the echo context to the parent context
+package echocontext
