+++ internal/httpserve/middleware/echocontext/errors.go
@@ -1,4 +1,4 @@
-package echomw
+package echocontext
 
 import (
 	"errors"
