--- internal/httpserve/middleware/echomw/doc.go
@@ -1,2 +1,0 @@
-// Package echomw adds middelware specific to echo
-package echomw
